Refactor OpenRouter client structure
- Moved OpenRouter client logic to a new package.
- Split functionality into multiple files for better organization.
- Introduced `client.go` for main client operations.
- Added `config.go` to handle configuration and validation.
- Created `messages.go` for message preparation and handling.
- Implemented request handling in `requests.go`.
- Managed streaming responses in `streaming.go`.
- Defined types in `types.go`.
