+++ internal/ai/openrouter/types.go
@@ -1,0 +1,25 @@
+package openrouter
+
+// ChatMessage represents a message in the conversation
+type ChatMessage struct {
+	Role    string `json:"role"`
+	Content string `json:"content"`
+}
+
+// streamResponse holds the structure for parsing streaming responses
+type streamResponse struct {
+	Choices []struct {
+		Delta struct {
+			Content string `json:"content"`
+		} `json:"delta"`
+	} `json:"choices"`
+}
+
+// completionResponse holds the structure for parsing completion responses
+type completionResponse struct {
+	Choices []struct {
+		Message struct {
+			Content string `json:"content"`
+		} `json:"message"`
+	} `json:"choices"`
+}
